🐛 (auth.service.ts): add missing async keyword to invalidateHackerToken method to fix asynchronous behavior 🐛 (categories.controller.ts): add missing import statements for HttpException, Inject, and CACHE_MANAGER 🐛 (categories.controller.ts): add missing constructor parameters for authService and cacheManager 🐛 (categories.controller.ts): add missing cache check and cache set logic to prevent unnecessary calls to invalidateHackerToken method 🐛 (categories.service.ts): add missing import statement for HttpException 🐛 (categories.service.ts): add check for categories on active menu before removing a category 🐛 (file-uploader.service.ts): remove unnecessary console.log statement 🐛 (menu.controller.ts): add missing import statements for HttpException and AuthService 🐛 (menu.controller.ts): add missing constructor parameter for authService 🐛 (menu.controller.ts): add missing cache check and cache set logic to prevent unnecessary calls to invalidateHackerToken method 🐛 (menu.service.ts): remove unnecessary debugger statement
📝 (products.controller.ts): Add imports for Inject, HttpException, Cache, AuthService, and CACHE_MANAGER to improve code readability and maintainability
📝 (products.controller.ts): Inject CACHE_MANAGER and AuthService in the constructor to enable caching and authentication functionality
📝 (products.controller.ts): Add cache check and invalidate hacker token logic to improve security and prevent unauthorized access
📝 (products.controller.ts): Add cache check and set cache logic to improve performance and reduce database queries
📝 (products.controller.ts): Add error handling for unauthorized user and throw HttpException with FORBIDDEN status
📝 (products.module.ts): Add imports for AuthModule and CacheModule to enable authentication and caching functionality
📝 (products.module.ts): Register CacheModule with ttl and max values to configure cache settings
📝 (products.service.ts): Add import for HttpException and HttpStatus to improve code readability and maintainability
📝 (products.service.ts): Add logic to check if product is associated with an active menu and throw HttpException if true
📝 (products.service.ts): Add logic to remove product from category_products table and throw HttpException if product is associated with an active menu
📝 (users.service.ts): Add method unsubscribeOnDb to update is_subscribed field in users table to false

+++ src/products/products.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { v4 as uuid } from 'uuid';
 import { Client } from 'pg';
 import { InjectConnection } from 'nest-postgres';
@@ -197,6 +197,24 @@ export class ProductsService {
   }
 
   async remove(id: string) {
+    const itemsOnActiveMenu = await this.dbConnection.query(
+      `select * from ${this.tenant}.menus m 
+      inner join ${this.tenant}.menu_categories mc on m.id = mc.menu_id 
+      inner join ${this.tenant}.category_products cp on cp.category_id = mc.category_id 
+      where cp.product_id = '${id}' and m.is_active = true`,
+    );
+
+    if (itemsOnActiveMenu?.rows?.length > 0) {
+      throw new HttpException(
+        'Produto cadastrado em um menu ativo.',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
+
+    await this.dbConnection.query(
+      `delete from ${this.tenant}.category_products where product_id = '${id}'`,
+    );
+
     const data = await this.dbConnection.query(
       `delete from ${this.tenant}.products where id = '${id}' returning *`,
     );
